Add several casts - a try to solve issue #1831.
Most of them are unneccessary, I know. But I have no
idea how to fix it otherwise. Now %li is casted to
(long) mostly everywhere.

+++ src/MaximaManual.cpp
@@ -276,7 +276,7 @@ void MaximaManual::CompileHelpFileAnchors(wxString maximaHtmlDir,
       {
         wxLogMessage(_("Have only %li keyword anchors at the end of parsing the maxima manual => "
                        "Not caching the result of using the built-in keyword list"),
-                     foundAnchorsTotal);
+                     (long)foundAnchorsTotal);
         LoadBuiltInManualAnchors();
       }
     else
